requirements.txt
fastapi
uvicorn
sqlalchemy
pyjwt
python-multipart
pytest
pytest-asyncio
httpx